(#1299) Removed collection's Filtered
