+++ src/test/java/org/cactoos/iterable/FilteredTest.java
@@ -23,16 +23,22 @@
  */
 package org.cactoos.iterable;
 
+import org.cactoos.list.ListOf;
 import org.cactoos.scalar.LengthOf;
 import org.hamcrest.MatcherAssert;
 import org.hamcrest.Matchers;
+import org.hamcrest.collection.IsEmptyIterable;
+import org.hamcrest.core.IsNot;
 import org.junit.Test;
+import org.llorllale.cactoos.matchers.Assertion;
+import org.llorllale.cactoos.matchers.ScalarHasValue;
 
 /**
  * Test case for {@link Filtered}.
  * @since 0.1
  * @checkstyle JavadocMethodCheck (500 lines)
  * @checkstyle MagicNumberCheck (500 lines)
+ * @checkstyle ClassDataAbstractionCouplingCheck (500 lines)
  */
 public final class FilteredTest {
 
@@ -80,6 +86,56 @@ public final class FilteredTest {
             "Can't filter the iterable twice, again",
             list, Matchers.iterableWithSize(3)
         );
+    }
+
+    @Test
+    public void filterEmptyList() {
+        new Assertion<>(
+            "Filter must work on empty collection",
+            new Filtered<String>(
+                input -> input.length() > 4,
+                new ListOf<>()
+            ),
+            new IsEmptyIterable<>()
+        ).affirm();
+    }
+
+    @Test
+    public void length() {
+        new Assertion<>(
+            "Size must be equal to number of items matching the filter",
+            new LengthOf(
+                new Filtered<>(
+                    input -> input.length() >= 4,
+                    new IterableOf<>("some", "text", "yes")
+                )
+            ),
+            new ScalarHasValue<>(2.)
+        ).affirm();
+    }
+
+    @Test
+    public void withItemsNotEmpty() {
+        new Assertion<>(
+            "Must not be empty with items",
+            new Filtered<String>(
+                input -> input.length() > 4,
+                new IterableOf<>("first", "second")
+            ),
+            new IsNot<>(new IsEmptyIterable<>())
+        ).affirm();
+    }
+
+    @Test
+    public void withoutItemsIsEmpty() {
+        new Assertion<>(
+            "Must be empty without items",
+            new Filtered<String>(
+                input -> input.length() > 16,
+                new IterableOf<>("third", "fourth")
+            ),
+            new IsEmptyIterable<>()
+        ).affirm();
     }
 
 }
